pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kie"
version = "3.0.0"
description = "Kearney Insight Engine v3 - AI-powered consulting delivery platform with KDS compliance"
readme = "README.md"
license = {text = "Proprietary"}
requires-python = ">=3.11"
authors = [
    {name = "Kearney Digital & Analytics"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    # Data processing
    "pandas>=2.1.0",
    "polars>=0.19.0",
    "numpy>=1.24.0",
    "openpyxl>=3.1.0",
    "xlsxwriter>=3.1.0",

    # Document generation
    "python-pptx>=0.6.21",
    "python-docx>=1.1.0",

    # API
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pydantic>=2.5.0",

    # Configuration
    "pyyaml>=6.0",

    # Utilities
    "rich>=13.0.0",
    "click>=8.1.0",
    "requests>=2.31.0",
    "aiohttp>=3.9.0",
    "tenacity>=8.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.7.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pre-commit>=3.5.0",
]
geo = [
    # Geospatial processing
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "folium>=0.15.0",
    "geopy>=2.4.0",
    "fiona>=1.9.0",
    # Optional paid geocoding services
    "googlemaps>=4.10.0",
    "mapbox>=0.18.0",
]
api = [
    # API enhancements
    "redis>=5.0.0",
    "celery>=5.3.0",
]
webapp = [
    "streamlit>=1.28.0",
]
all = [
    "kie[dev,geo,api,webapp]",
]

[project.scripts]
kie = "kie.cli:main"

[project.urls]
Homepage = "https://github.com/kearney/kie"
Documentation = "https://github.com/kearney/kie/docs"

[tool.setuptools.packages.find]
where = ["."]
include = ["kie*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
filterwarnings = [
    "ignore::DeprecationWarning",
]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
strict = false

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["kie"]

[tool.black]
line-length = 100
target-version = ["py311"]
